requirements.txt
requests>=2.31
beautifulsoup4>=4.12
psycopg2-binary>=2.9
playwright>=1.40
flask>=3.0
